refactor Card to not require offset (use CSS style instead), remove reundandant Card2
