+++ app/components/Card/index.js
@@ -3,18 +3,11 @@
  */
 import React from 'react';
 import { VectorCards } from 'ab-vector-cards';
-import styled from 'styled-components';
+
+import { CardBack, CardFront, CardStyle } from '../Seat/styles';
 
 const values = ['2', '3', '4', '5', '6', '7', '8', '9', '10', 'j', 'q', 'k', 'a'];
 const suits = ['clubs', 'diamonds', 'hearts', 'spades'];
-
-const CardWrapper = styled.div`
-  position: relative;
-  margin-left: ${(props) => props.offset[0]}%;
-  margin-top: ${(props) => props.offset[1]}%;
-  max-width: 100%;
-  height: auto;
-`;
 
 function Card(props) {
   const vc = new VectorCards();
@@ -25,24 +18,35 @@ function Card(props) {
   if (props.cardNumber >= 0) {
     link = vc.getCardData(props.size, suit, value);
   } else {
-    link = vc.getBackData(props.size, '#7A7BB8', '#2E319C');
+    link = vc.getBackData(props.size, '#32B7D3', '#217C8F');
   }
 
   // Note: meaning of card numbers
   //  * -1 stands for back side of cards,
   //  * null stands for no card
   //  * > 0  stands for normal cards
-  if (!props.folded && props.cardNumber !== null) {
+  if (!props.folded && props.cardNumber === -1) {
     return (
-      <CardWrapper offset={props.offset}>
-        <img
+      // TODO: remove offset prop and from CardShared style?
+      <CardBack>
+        <CardStyle
           key={suit + value}
           src={link}
-          className="card"
           alt=""
-          style={{ maxWidth: '100%', height: 'auto' }}
         />
-      </CardWrapper>
+      </CardBack>
+    );
+  }
+  if (!props.folded && props.cardNumber !== null) {
+    return (
+      // TODO: remove offset prop and from CardShared style?
+      <CardFront>
+        <CardStyle
+          key={suit + value}
+          src={link}
+          alt=""
+        />
+      </CardFront>
     );
   }
   return null;
@@ -50,7 +54,6 @@ function Card(props) {
 
 Card.propTypes = {
   cardNumber: React.PropTypes.number,
-  offset: React.PropTypes.array,
   size: React.PropTypes.number,
   folded: React.PropTypes.bool,
 };
